Resolved NSMakeRect() argument casting TODO in sizing_darwin.m.

+++ redo/sizing_darwin.m
@@ -8,6 +8,5 @@
 
 void moveControl(id c, intptr_t x, intptr_t y, intptr_t width, intptr_t height)
 {
-	// TODO appropriate casts
-	[toNSView(c) setFrame:NSMakeRect(x, y, width, height)];
+	[toNSView(c) setFrame:NSMakeRect((CGFloat) x, (CGFloat) y, (CGFloat) width, (CGFloat) height)];
 }
